move generic geometry types to PathData
